functions/requirements.txt
# Firebase Functions Python Dependencies
functions-framework==3.*
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0

# Web Framework
flask==2.2.3
werkzeug==2.2.3

# Payment Processing
stripe==7.0.0
paypalrestsdk==1.13.1

# PDF Processing
xhtml2pdf==0.2.11
pdfplumber==0.9.0
PyPDF2==3.0.1
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.26.4
opencv-python==4.8.0.74

# Utilities
requests==2.31.0
python-dateutil==2.8.2
cryptography==41.0.7
python-dotenv==1.0.0